Bump packages to v0.1.12 and use npm pkg set for script updates
Updated all package versions from 0.1.11 to 0.1.12. The main change is in the setup package, which now uses `npm pkg set` to modify package.json scripts instead of manually writing the file. This prevents the loss of newly installed dependencies during the setup process. Also includes some internal renaming for consistency.

The code now has a 100ms busy wait to 'ensure file system sync' because apparently we don't trust the OS to do its one job anymore. It's like waiting for your toast to pop up, but instead of toast, it's just the faint hope that the file system isn't lying to you.